Safari-609.1.11.1/LayoutTests/platform/ios-wk2/editing/selection/5825350-2-expected.txt

CONSOLE MESSAGE: line 14: TypeError: undefined is not a function (evaluating 'textInputController.selectedRange()')
This tests for a bug where moving the caret right towards a non-editable pocket of an editable region would make the caret disappear. The caret should be just after the 'y' in "Sally".
Bob ; Sally
